Iniciado ingestor: problemas de codificación de las páginas
