+++ kapanlagi/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "ingestor-interesting",
+    "name": "ingestor-kapanlagi",
     "version": "0.0.1",
     "license": "UNLICENSED",
     "dependencies": {
