+++ kapanlagi/template.js
@@ -2,7 +2,11 @@
 
 const structure_template = (`
 <section class="post-heading">
-    {{{ post_heading }}}
+    <h1> {{title}} </h1>
+    <div> {{{metadata}}} </div>
+    <div class="main-image">
+        <img data-libingester-asset-id="{{ main_image.asset_id }}">
+    </div>
 </section>
 <section class="post-body">
     {{{ post_body }}}
